feat(): abstract BeepConfig BeepPreview and BeepConfirm
