+++ src/app/mint/beep/components/mint_button.tsx
@@ -25,26 +25,6 @@ const MintButton = () => {
     )
   }
 
-  // const renderClaimNFTButton = () => (
-  //   <Web3Button
-  //     action={contract => contract.erc721.claim(1)}
-  //     className="!h-12 !w-full !bg-black !text-xl !rounded-full !transition-colors hover:!bg-[#1F1F1F] [&>svg>circle]:!stroke-white"
-  //     contractAddress={env.NEXT_PUBLIC_BEEP_CONTRACT_ADDRESS}
-  //     onError={error => {
-  //       toast.error((error as unknown as { reason: string })?.reason || 'Failed to claim NFT')
-  //     }}
-  //     onSuccess={data => {
-  //       const tokenId = data[0].id.toNumber()
-  //       setTokenId(tokenId.toString())
-  //       onOpen()
-  //     }}
-  //     theme="light"
-  //   >
-
-  //     Mint now
-  //   </Web3Button>
-  // )
-
   return (
     <div className="mt-4">
       <DeployModal
